[Important] Enforce light theme on Windows native project
Because Telerik UI for MAUI doesn't have full dark/light theme variant support yet, we want to make sure that the windows native platform requests Light them on the user's computer.

This important because many users have mixed mode or full dark mode on their windows 10 and 11 systems, which will break the appearance of the app.